requirements.txt
librosa
pandas
numpy
matplotlib
seaborn
scikit-learn
openai-whisper
torch
torchvision